* Changed WBFS stuff to C++ classes * Fixed bug in http.c (where realloc *might* choose another address, thanks Dr. Clipper) * Added support to /wbfs/Game Title [GAMEID].wbfs files (thanks oggzee) * Fixed bug in cfg_cleanup when switching partitions * Added BGM class again (for playing background music, requested by dimok) * Added settings for background music again * Fixed bug in MEM2 class (returning an invalid handle when no memory could be allocated, thanks to dimok) * Updated DIP module to OpenDIP (report bugs with this one if you've found them, but report them here: http://github.com/spacemanspiff/odip-plugin) * Added initial code for cios 222 rev5 (THIS DOES NOT WORK YET! DON'T FILE BUGS ON THIS ONE!) * Added fatffs module by Hermes and Waninkoko (THIS DOES NOT WORK YET! DON'T FILE BUGS ON THIS ONE!) * Fixed bug in Settings, which resulted in a crash when the partition was changed. * Added caching for gamelist entries, so switching between different screens/sort options/display modes should be faster * Changed defines in ehc_module, in order to prevent clashes with new defines in rev5

+++ source/usbloader/partition_usbloader.h
@@ -38,32 +38,32 @@ typedef struct {
 
 typedef struct
 {
-	int fs_type;
-	int part_fs;
-	int wbfs_i;  // seq wbfs part index
-	int fat_i;   // seq fat part index
-	int ntfs_i;  // seq ntfs part index
-	int index;
+	u8 fs_type;
+	u8 part_fs;
+	u8 wbfs_i;  // seq wbfs part index
+	u8 fat_i;   // seq fat part index
+	u8 ntfs_i;  // seq ntfs part index
+	u8 index;
 } PartInfo;
 
 typedef struct
 {
-	int num;
+	u8 num;
 	u32 sector_size;
 	partitionEntry pentry[MAX_PARTITIONS_EX];
-	int wbfs_n;
-	int fat_n;
-	int ntfs_n;
+	u8 wbfs_n;
+	u8 fat_n;
+	u8 ntfs_n;
 	PartInfo pinfo[MAX_PARTITIONS_EX];
 } PartList;
 
 /* Prototypes */
 s32 Partition_GetEntries(u32 device, partitionEntry *outbuf, u32 *outval);
-s32 Partition_GetEntriesEx(u32 device, partitionEntry *outbuf, u32 *outval, int *num);
+s32 Partition_GetEntriesEx(u32 device, partitionEntry *outbuf, u32 *outval, u8 *num);
 bool Device_ReadSectors(u32 device, u32 sector, u32 count, void *buffer);
 bool Device_WriteSectors(u32 device, u32 sector, u32 count, void *buffer);
 s32 Partition_GetList(u32 device, PartList *plist);
-int Partition_FixEXT(u32 device, int part);
+int Partition_FixEXT(u32 device, u8 part);
 
 bool  part_is_extended(int type);
 bool  part_is_data(int type);
